@@ -10,23 +10,28 @@
     "docs:build": "vuepress build docs"
   },
   "dependencies": {
+    "@ant-design/icons-vue": "^5.1.5",
+    "ant-design-vue": "^2.0.0-beta.9",
+    "axios": "^0.20.0",
     "core-js": "^2.6.5",
-    "vue": "^2.6.10",
-    "vue-router": "^3.1.5",
+    "mockjs": "^1.1.0",
+    "vue": "^3.0.0-beta.1",
+    "vue-router": "4.0.0-beta.9",
     "vuepress-theme-reco": "^1.2.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.12.0",
     "@vue/cli-plugin-eslint": "^3.12.0",
     "@vue/cli-service": "^3.12.0",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "@vue/eslint-config-standard": "^4.0.0",
     "babel-eslint": "^10.0.1",
     "cz-conventional-changelog": "^3.1.0",
     "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
     "node-sass": "^4.13.1",
     "typescript": "^3.8.3",
-    "vue-template-compiler": "^2.6.10",
+    "vue-cli-plugin-vue-next": "~0.1.3",
     "vuepress": "^1.1.0",
     "vuepress-plugin-flowchart": "^1.4.3",
     "webpack": "^4.41.6",
@@ -38,7 +43,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "@vue/standard"
     ],
     "rules": {},
