@@ -2,7 +2,10 @@
  * @author 青夏
  * @version 3.0
  * @description
- * 这是vuepress侧边栏目录生成辅助的V3版本，V3相比与V2来说，使用了TS编写，简化了调用方法，只需要一次配置，即可自动生成一个分栏下的所有分类，使用方法：调用start函数，传入你需要自动生成目录的文件夹名，比如需要生成mood下的侧边栏目录，你只需要调用start('mood')即可，同时，V3版本将每个分页的标题数据写在了对应目录的json文件下，这样你需要配置每个分栏的名字时，只需要在对应目录下创建一个json文件，写入对应数据即可，同时，引入了json配置文件的思想后，你可以对本插件进行更多的功能拓展，比如配置排序等等需求
+ * 这是vuepress侧边栏目录生成辅助的V3版本，V3相比与V2来说，使用了TS编写，简化了调用方法，只需要一次配置，即可自动生成一个分栏下的所有分类.
+ * 使用方法：调用start函数，传入你需要自动生成目录的文件夹名，比如需要生成mood下的侧边栏目录，你只需要调用start('mood')即可，
+ * 同时，V3版本将每个分页的标题数据写在了对应目录的json文件下，这样你需要配置每个分栏的名字时，只需要在对应目录下创建一个json文件，写入对应数据即可，
+ * 引入了json配置文件的思想后，你可以对本插件进行更多的功能拓展，比如配置排序等等需求
  *
  *
  */
